pyproject.toml
# Specify the build system.
[build-system]
requires = ["sip >=6,<7", "PyQt-builder >=1.17,<2", "PyQt6==6.8.1", "ruff"]
build-backend = "sipbuild.api"

# Specify the PEP 566 metadata for the project.
[project]
name = "PyQt6Qlementine"
version = "1.4.2"
requires-python = ">=3.10"
description = "Python bindings for Qlementine, a modern QStyle for desktop Qt6 applications"
license = { text = "MIT" }
readme = "README.md"
dependencies = ["PyQt6>=6.8.0"]
classifiers = [
	"Development Status :: 3 - Alpha",
	"Intended Audience :: Developers",
	"License :: OSI Approved :: MIT License",
	"Operating System :: MacOS",
	"Operating System :: Microsoft :: Windows",
	"Operating System :: POSIX :: Linux",
	"Programming Language :: Python :: 3",
	"Topic :: Software Development :: Libraries :: Python Modules",
	"Topic :: Software Development :: Widget Sets",
	"Environment :: X11 Applications :: Qt",
	"Typing :: Typed",
]

[project.urls]
Upstream = "https://github.com/oclero/qlementine"
Homepage = "https://github.com/pyapp-kit/PyQt6Qlementine"

[dependency-groups]
build = ["sip >=6,<7", "PyQt-builder >=1.17,<2", "PyQt6==6.8.1", "PyQt6-Qt6==6.8.2"]
test = ["pytest>=8.3.4"]
dev = [
	{ "include-group" = "build" },
	{ "include-group" = "test" },
	"ruff>=0.9.8",
	"aqtinstall>=3.2.0",
	"prek>=0.3.1",
]

[tool.sip.builder]
jobs = 8

[tool.sip.bindings.PyQt6Qlementine]
pep484-pyi = true
sip-file = "PyQt6Qlementine.sip"
qmake-QT = ["widgets", "svg"]
include-dirs = [
	"qlementine/lib/include",
	"qlementine/lib/src",
]
headers = [
	"qlementine/lib/include/oclero/qlementine/Common.hpp",
	"qlementine/lib/include/oclero/qlementine/animation/WidgetAnimationManager.hpp",
	"qlementine/lib/include/oclero/qlementine/animation/WidgetAnimator.hpp",
	"qlementine/lib/include/oclero/qlementine/style/Delegates.hpp",
	"qlementine/lib/include/oclero/qlementine/style/QlementineStyle.hpp",
	"qlementine/lib/include/oclero/qlementine/style/QlementineStyleOption.hpp",
	"qlementine/lib/include/oclero/qlementine/style/Theme.hpp",
	"qlementine/lib/include/oclero/qlementine/style/ThemeManager.hpp",
	"qlementine/lib/include/oclero/qlementine/tools/ThemeEditor.hpp",
	"qlementine/lib/include/oclero/qlementine/widgets/AboutDialog.hpp",
	"qlementine/lib/include/oclero/qlementine/widgets/AbstractItemListWidget.hpp",
	"qlementine/lib/include/oclero/qlementine/widgets/Action.hpp",
	"qlementine/lib/include/oclero/qlementine/widgets/ActionButton.hpp",
	"qlementine/lib/include/oclero/qlementine/widgets/ColorButton.hpp",
	"qlementine/lib/include/oclero/qlementine/widgets/ColorEditor.hpp",
	"qlementine/lib/include/oclero/qlementine/widgets/CommandLinkButton.hpp",
	"qlementine/lib/include/oclero/qlementine/widgets/Expander.hpp",
	"qlementine/lib/include/oclero/qlementine/widgets/IconWidget.hpp",
	"qlementine/lib/include/oclero/qlementine/widgets/Label.hpp",
	"qlementine/lib/include/oclero/qlementine/widgets/LineEdit.hpp",
	"qlementine/lib/include/oclero/qlementine/widgets/LoadingSpinner.hpp",
	"qlementine/lib/include/oclero/qlementine/widgets/Menu.hpp",
	"qlementine/lib/include/oclero/qlementine/widgets/NavigationBar.hpp",
	"qlementine/lib/include/oclero/qlementine/widgets/NotificationBadge.hpp",
	"qlementine/lib/include/oclero/qlementine/widgets/PlainTextEdit.hpp",
	"qlementine/lib/include/oclero/qlementine/widgets/Popover.hpp",
	"qlementine/lib/include/oclero/qlementine/widgets/PopoverButton.hpp",
	"qlementine/lib/include/oclero/qlementine/widgets/RoundedFocusFrame.hpp",
	"qlementine/lib/include/oclero/qlementine/widgets/SegmentedControl.hpp",
	"qlementine/lib/include/oclero/qlementine/widgets/StatusBadgeWidget.hpp",
	"qlementine/lib/include/oclero/qlementine/widgets/Switch.hpp",
	# internal header needing moc (has Q_OBJECT)
	"qlementine/lib/src/style/eventFilters/WidgetWithFocusFrameEventFilter.hpp",
]
sources = [
	"qlementine/lib/src/animation/WidgetAnimationManager.cpp",
	"qlementine/lib/src/animation/WidgetAnimator.cpp",
	"qlementine/lib/src/resources/ResourceInitialization.cpp",
	"qlementine/lib/src/style/Delegates.cpp",
	"src/EventFilters_wrapper.cpp",
	"qlementine/lib/src/style/QlementineStyle.cpp",
	"qlementine/lib/src/style/Theme.cpp",
	"qlementine/lib/src/style/ThemeManager.cpp",
	"qlementine/lib/src/tools/ThemeEditor.cpp",
	"qlementine/lib/src/utils/BadgeUtils.cpp",
	"qlementine/lib/src/utils/ColorUtils.cpp",
	"qlementine/lib/src/utils/FontUtils.cpp",
	"qlementine/lib/src/utils/GeometryUtils.cpp",
	"qlementine/lib/src/utils/IconUtils.cpp",
	"qlementine/lib/src/utils/ImageUtils.cpp",
	"qlementine/lib/src/utils/LayoutUtils.cpp",
	"qlementine/lib/src/utils/MenuUtils.cpp",
	"qlementine/lib/src/utils/PrimitiveUtils.cpp",
	"qlementine/lib/src/utils/RadiusesF.cpp",
	"qlementine/lib/src/utils/StateUtils.cpp",
	"qlementine/lib/src/utils/StyleUtils.cpp",
	"qlementine/lib/src/utils/WidgetUtils.cpp",
	"qlementine/lib/src/widgets/AboutDialog.cpp",
	"qlementine/lib/src/widgets/AbstractItemListWidget.cpp",
	"qlementine/lib/src/widgets/Action.cpp",
	"qlementine/lib/src/widgets/ActionButton.cpp",
	"qlementine/lib/src/widgets/ColorButton.cpp",
	"qlementine/lib/src/widgets/ColorEditor.cpp",
	"qlementine/lib/src/widgets/CommandLinkButton.cpp",
	"qlementine/lib/src/widgets/Expander.cpp",
	"qlementine/lib/src/widgets/IconWidget.cpp",
	"qlementine/lib/src/widgets/Label.cpp",
	"qlementine/lib/src/widgets/LineEdit.cpp",
	"qlementine/lib/src/widgets/LoadingSpinner.cpp",
	"qlementine/lib/src/widgets/Menu.cpp",
	"qlementine/lib/src/widgets/NavigationBar.cpp",
	"qlementine/lib/src/widgets/NotificationBadge.cpp",
	"qlementine/lib/src/widgets/PlainTextEdit.cpp",
	"src/Popover_patched.cpp",
	"qlementine/lib/src/widgets/PopoverButton.cpp",
	"qlementine/lib/src/widgets/RoundedFocusFrame.cpp",
	"qlementine/lib/src/widgets/SegmentedControl.cpp",
	"qlementine/lib/src/widgets/StatusBadgeWidget.cpp",
	"qlementine/lib/src/widgets/Switch.cpp",
]


[tool.cibuildwheel]
build-verbosity = 1
build = [
	"cp310-win_amd64",
	"cp310-manylinux_x86_64",
	"cp310-macosx_x86_64",
	"cp310-macosx_arm64",
]
manylinux-x86_64-image = "manylinux_2_28"
build-frontend = "build[uv]"
test-command = "pytest {project}/tests -v"
test-groups = ["test"]

[tool.cibuildwheel.linux]
environment-pass = ["QT_VERSION"]
before-build = [
	"yum install -y libxkbcommon-devel",
	"uvx --from aqtinstall==3.2.0 aqt install-qt linux desktop $QT_VERSION --outputdir Qt --base http://mirrors.ocf.berkeley.edu/qt/",
]
repair-wheel-command = ["python scripts/repair_wheel.py {dest_dir} {wheel}"]
test-command = [
	"yum install -y epel-release",
	"yum install -y libxkbcommon-x11 xcb-util-cursor xcb-util-wm xcb-util-keysyms xorg-x11-server-Xvfb",
	"uv pip install pytest-xvfb",
	"pytest {project}/tests -v",
]

[tool.cibuildwheel.macos]
before-build = "uvx --from aqtinstall==3.2.0 aqt install-qt mac desktop $QT_VERSION --outputdir Qt --base http://mirrors.ocf.berkeley.edu/qt/"
repair-wheel-command = ["python scripts/repair_wheel.py {dest_dir} {wheel}"]

[tool.cibuildwheel.windows]
before-build = "uvx --from aqtinstall==3.2.0 aqt install-qt windows desktop %QT_VERSION% win64_msvc2019_64 --outputdir Qt --base http://mirrors.ocf.berkeley.edu/qt/"


# https://docs.astral.sh/ruff/
[tool.ruff]
line-length = 88
target-version = "py39"
src = ["src", "tests"]
fix = true
unsafe-fixes = true

[tool.ruff.lint]
pydocstyle = { convention = "numpy" }
select = [
	"E",   # style errors
	"F",   # flakes
	"W",   # warnings
	"I",   # isort
	"UP",  # pyupgrade
	"TC",  # flake8-typecheck
	"TID", # flake8-tidy-imports
	"RUF", # ruff-specific rules
]

[tool.ruff.lint.per-file-ignores]
"tests/*.py" = ["RUF012"]

# https://docs.astral.sh/ruff/formatter/
[tool.ruff.format]
docstring-code-format = true
